Allow container to be run as non-root
Make changes to the Dockerfile and package scripts to make image
amenable to running as non-root and with a read-only root file
system, both security best practices.

Move workspace out from under root home directory so it is accessible
to non-root users.  Provide package script for Docker container that
does not remove the build directory since this is not present in the
Docker container and trying to remove it fails for non-root users.

Also add EXPOSE command to Dockerfile to help people figure out what
port this app listens on.